+++ src/components/layout/footer.tsx
@@ -1,3 +1,9 @@
 export const Footer = () => {
-  return <footer className={'w-full h-5 p-2 bg-gray-300'}></footer>;
+  return (
+    <footer
+      className={
+        'w-full h-2 p-2 bg-gray-100 bg-gradient-to-b from-gray-100 to-blue-200'
+      }
+    />
+  );
 };